+++ src/http.ts
@@ -1,17 +1,28 @@
 import { ServerResponse } from 'http';
 
+// INFORMATION
+
+/** The client should continue the request or ignore the response if the request is already finished. */
+export const HTTP_100_Continue = 100;
+/** Response to an Upgrade request header from the client and indicates the protocol the server is switching to. */
+export const HTTP_101_SwitchingProtocols = 101;
+
+// SUCCESS
+
+/** The request succeeded, and the response body contains the requested resource. */
 export const HTTP_200_OK = 200;
-/**
- * The request succeeded, and a new resource was created as a result.
- * This is typically the response sent after POST requests,
- * or some PUT requests.
- **/
+/** The request succeeded, and a new resource was created as a result. This is typically the response sent after POST or PUT requests. */
 export const HTTP_201_Created = 201;
-/*
- * The request has been received but not yet acted upon.
- * Uee in cases where another process or server handles the request.
- */
+/* The request has been received but not yet acted upon. Another process or server handles the request. */
 export const HTTP_202_Accepted = 202;
+/** The returned metadata is not necessarily complete. */
+export const HTTP_203_NonAuthoritativeInformation = 203;
+/** The response body is empty. */
+export const HTTP_204_NoContent = 204;
+/** The request succeeded, but the returned metadata is not necessarily complete. */
+export const HTTP_206_PartialContent = 206;
+
+// REDIRECTION
 
 /**
  * Only safe to use in response to GET and HEAD requests
@@ -32,26 +43,82 @@ export const HTTP_304_NotModified = 304;
 export const HTTP_307_TemporaryRedirect = 307;
 export const HTTP_308_PermanentRedirect = 308;
 
-/** The request is malformed (e.g. a URL param is of a wrong type) */
+// CLIENT ERROR
+
+/** The request is malformed (e.g. a URL param is of a wrong type). */
 export const HTTP_400_BadRequest = 400;
-/** User is not authenticated (i.e. not logged in) */
+/** User is not authenticated (i.e. not logged in). */
 export const HTTP_401_Unauthorized = 401;
-/** User is logged in but doesn't have the necessary privileges */
+/** User is logged in but doesn't have the necessary privileges. */
 export const HTTP_403_Forbidden = 403;
-/** The request looks OK but the resource does not exist */
+/** The request looks OK but the resource does not exist. */
 export const HTTP_404_NotFound = 404;
-/**
- * The resource has been permanently deleted from server, with no forwarding
- * address.
- */
+/** The request method is not supported by the target resource. */
+export const HTTP_405_MethodNotAllowed = 405;
+/** The server can't produce a response matching the list of acceptable types. */
+export const HTTP_406_NotAcceptable = 406;
+/** Similar to `HTTP_401_Unauthorized` but authentication is needed to be done by a proxy. */
+export const HTTP_407_ProxyAuthenticationRequired = 407;
+/** The server timed out waiting for the request. */
+export const HTTP_408_RequestTimeout = 408;
+/** The request conflicts with the current state of the server. */
+export const HTTP_409_Conflict = 409;
+/** The resource has been permanently deleted from server, with no forwarding address. */
 export const HTTP_410_Gone = 410;
+/** The request's Content-Length header field is not defined and the server requires it. */
+export const HTTP_411_LengthRequired = 411;
+/** The client has indicated preconditions in its headers which the server does not meet. */
+export const HTTP_412_PreconditionFailed = 412;
+/** The request is larger than the server is willing or able to process. */
+export const HTTP_413_PayloadTooLarge = 413;
+/** The URI requested by the client is longer than the server is willing to interpret. */
+export const HTTP_414_URITooLong = 414;
+/** The media format of the requested data is not supported by the server */
+export const HTTP_415_UnsupportedMediaType = 415;
+/** The Range header field in the request cannot be fulfilled. It's possible that the range is outside the size of the target URI's data. */
+export const HTTP_416_RangeNotSatisfiable = 416;
+/** The expectation indicated by the Expect request header field cannot be met by the server. */
+export const HTTP_417_ExpectationFailed = 417;
 /** The server refuses the attempt to brew coffee with a teapot. */
 export const HTTP_418_ImATeapot = 418;
+/** The request was directed at a server that is not able to produce a response. */
+export const HTTP_421_MisdirectedRequest = 421;
+/** The server refuses to perform the request using the current protocol but might be willing to do so after the client upgrades to a different protocol. */
+export const HTTP_426_UpgradeRequired = 426;
+/** The origin server requires the request to be conditional. */
+export const HTTP_428_PreconditionRequired = 428;
+/** Received too many requests in a given amount of time ("rate limiting"). */
+export const HTTP_429_TooManyRequests = 429;
+/** The server is unwilling to process the request because its header fields are too large. */
+export const HTTP_431_RequestHeaderFieldsTooLarge = 431;
+/** The resource cannot legally be provided, such as a web page censored by a government. */
+export const HTTP_451_UnavailableForLegalReasons = 451;
 
+// SERVER ERROR
+
+/** The server has encountered a situation it does not know how to handle. */
 export const HTTP_500_InternalServerError = 500;
+/** The request method type is not supported by the server and cannot be handled. */
+export const HTTP_501_NotImplemented = 501;
+/** The server, while working as a gateway to get a response needed to handle the request, got an invalid response. */
+export const HTTP_502_BadGateway = 502;
+/** The server is not ready to handle the request. (Commonly: down for maintenance or overloaded.) */
+export const HTTP_503_ServiceUnavailable = 503;
+/** The server is acting as a gateway and cannot get a response in time. */
+export const HTTP_504_GatewayTimeout = 504;
+/** The HTTP version used in the request is not supported by the server. */
+export const HTTP_505_HTTPVersionNotSupported = 505;
+/** The server has an internal configuration error: the chosen variant resource is configured to engage in transparent content negotiation itself. */
+export const HTTP_506_VariantAlsoNegotiates = 506;
+/** Further extensions to the request are required for the server to fulfill it. */
+export const HTTP_510_NotExtended = 510;
+/** The client needs to authenticate to gain network access. */
+export const HTTP_511_NetworkAuthenticationRequired = 511;
 
 // ---------------------------------------------------------------------------
-// Types for HTTP Status codes
+// Union Types for the more commonly occurrring HTTP Status codes
+
+export type HTTP_INFO = typeof HTTP_100_Continue | typeof HTTP_101_SwitchingProtocols;
 
 export type HTTP_SUCCESS =
   | typeof HTTP_200_OK
@@ -65,39 +132,67 @@ export type HTTP_REDIRECTION =
   | typeof HTTP_304_NotModified
   | typeof HTTP_307_TemporaryRedirect
   | typeof HTTP_308_PermanentRedirect;
-
 export type HTTP_NOTMODIFIED = typeof HTTP_304_NotModified;
-
-export type HTTP_CLIENT_ERROR =
-  | typeof HTTP_400_BadRequest
-  | typeof HTTP_401_Unauthorized
-  | typeof HTTP_403_Forbidden
-  | typeof HTTP_404_NotFound
-  | typeof HTTP_410_Gone;
-/*
-  NOPE: No joke statuses in this utilitarian type union.
-  Later we might like to create a wider union with literally all the HTTP_4xx_*
-  status codes, including these more esoteric ones.
-  `HTTP_CLIENT_ERROR_ALL`, `HTTP_SERVER_ERROR_ALL`, `HTTP_ERROR_ALL`, etc.
-*/
-// | typeof HTTP_418_ImATeapot;
 
 export type HTTP_NOT_FOUND =
   | typeof HTTP_400_BadRequest
   | typeof HTTP_404_NotFound
   | typeof HTTP_410_Gone;
+
 export type HTTP_BANNED = typeof HTTP_401_Unauthorized | typeof HTTP_403_Forbidden;
+
+export type HTTP_CLIENT_ERROR = HTTP_NOT_FOUND | HTTP_BANNED;
 
 export type HTTP_SERVER_ERROR = typeof HTTP_500_InternalServerError;
 
 export type HTTP_ERROR = HTTP_CLIENT_ERROR | HTTP_SERVER_ERROR;
 
 export type HTTP_STATUS =
+  | HTTP_INFO_ALL
+  | HTTP_SUCCESS_ALL
+  | HTTP_REDIRECTION_ALL
+  | HTTP_CLIENT_ERROR_ALL
+  | HTTP_SERVER_ERROR_ALL;
+
+// ---------------------------------------------------------------------------
+// All HTTP Status codes, even the esoteric ones
+
+export type HTTP_INFO_ALL = HTTP_INFO;
+export type HTTP_SUCCESS_ALL =
   | HTTP_SUCCESS
-  | HTTP_REDIRECTION
+  | typeof HTTP_203_NonAuthoritativeInformation
+  | typeof HTTP_204_NoContent
+  | typeof HTTP_206_PartialContent;
+export type HTTP_REDIRECTION_ALL = HTTP_REDIRECTION;
+export type HTTP_CLIENT_ERROR_ALL =
   | HTTP_CLIENT_ERROR
+  | typeof HTTP_405_MethodNotAllowed
+  | typeof HTTP_406_NotAcceptable
+  | typeof HTTP_407_ProxyAuthenticationRequired
+  | typeof HTTP_408_RequestTimeout
+  | typeof HTTP_409_Conflict
+  | typeof HTTP_411_LengthRequired
+  | typeof HTTP_412_PreconditionFailed
+  | typeof HTTP_413_PayloadTooLarge
+  | typeof HTTP_414_URITooLong
+  | typeof HTTP_415_UnsupportedMediaType
+  | typeof HTTP_416_RangeNotSatisfiable
+  | typeof HTTP_417_ExpectationFailed
   | typeof HTTP_418_ImATeapot
-  | HTTP_SERVER_ERROR;
+  | typeof HTTP_421_MisdirectedRequest
+  | typeof HTTP_426_UpgradeRequired
+  | typeof HTTP_428_PreconditionRequired;
+export type HTTP_SERVER_ERROR_ALL =
+  | HTTP_SERVER_ERROR
+  | typeof HTTP_501_NotImplemented
+  | typeof HTTP_502_BadGateway
+  | typeof HTTP_503_ServiceUnavailable
+  | typeof HTTP_504_GatewayTimeout
+  | typeof HTTP_505_HTTPVersionNotSupported
+  | typeof HTTP_506_VariantAlsoNegotiates
+  | typeof HTTP_510_NotExtended
+  | typeof HTTP_511_NetworkAuthenticationRequired;
+export type HTTP_ERROR_ALL = HTTP_CLIENT_ERROR_ALL | HTTP_SERVER_ERROR_ALL;
 
 // ---------------------------------------------------------------------------